refactor: move common types to a separate package
