+++ packages/core/tests/fixtures/class/namespace/decorators.ts
@@ -1,7 +1,7 @@
+import {DecorableClass} from '@ast-decorators/typings';
 import {template} from '@babel/core';
 import {NodePath} from '@babel/traverse';
 import {Statement} from '@babel/types';
-import {DecorableClass} from '../../../../src/utils';
 
 let count = 0;
