+++ packages/typings/lib/index.d.ts
@@ -1,0 +1,15 @@
+import {
+  ClassDeclaration,
+  ClassExpression,
+  ClassMethod,
+  ClassPrivateMethod,
+  ClassPrivateProperty,
+  ClassProperty,
+} from '@babel/types';
+
+export type DecorableClass = ClassDeclaration | ClassExpression;
+export type DecorableClassMember =
+  | ClassProperty
+  | ClassPrivateProperty
+  | ClassMethod
+  | ClassPrivateMethod;
