+++ packages/typings/package.json
@@ -1,17 +1,13 @@
 {
-  "name": "@ast-decorators/core",
+  "name": "@ast-decorators/typings",
   "version": "0.1.0",
   "description": "AST Decorators core plugin",
-  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
   "scripts": {
     "build": "npm run build:remove && run-p build:js build:dts",
     "build:js": "babel src -d lib -x .ts --config-file ../../babel.config.json",
     "build:dts": "tsc -p tsconfig.build.json --emitDeclarationOnly",
     "build:remove": "rimraf lib",
     "typecheck": "tsc -p tsconfig.json --noEmit"
-  },
-  "dependencies": {
-    "@ast-decorators/typings": "^0.1.0",
-    "@babel/types": "^7"
   }
 }
