+++ packages/typings/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig",
+  "compilerOptions": {
+    "declaration": true,
+    "outDir": "lib"
+  },
+  "include": ["src/**/*", "tests/**/*"]
+}
